requirements.txt
addict==2.4.0
aliyun-python-sdk-core==2.15.1
aliyun-python-sdk-kms==2.16.2
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
blinker==1.7.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
crcmod==1.7
cryptography==42.0.7
cycler==0.12.1
dash==2.16.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
docker-pycreds==0.4.0
easydict==1.13
einops==0.8.0
filelock==3.13.1
Flask==3.0.3
fonttools==4.47.2
fsspec==2023.12.2
gitdb==4.0.11
GitPython==3.1.41
huggingface-hub==0.21.4
hydra-core==1.3.2
idna==3.6
imageio==2.34.1
imageio-ffmpeg==0.5.1
importlib_metadata==7.1.0
itsdangerous==2.1.2
Jinja2==3.1.3
jmespath==0.10.0
joblib==1.3.2
kiwisolver==1.4.5
lazy_loader==0.4
lxml==5.2.2
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.4
matplotlib==3.8.4
mdurl==0.1.2
mmcv==2.2.0
mmengine==0.10.4
model-index==0.1.11
mplcursors==0.5.3
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.2.1
numpy==1.26.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
opencv-python==4.9.0.80
opendatalab==0.0.10
openmim==0.3.9
openxlab==0.0.38
ordered-set==4.1.0
oss2==2.17.0
packaging==23.2
pandas==2.2.1
pillow==10.2.0
platformdirs==4.2.1
plotly==5.20.0
protobuf==4.25.2
psutil==5.9.8
pycocotools==2.0.7
pycparser==2.22
pycryptodome==3.20.0
Pygments==2.18.0
pyparsing==3.1.1
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
python-dateutil==2.8.2
pytorch-sidu==1.1.3
pytz==2023.4
PyYAML==6.0.1
regex==2023.12.25
requests==2.28.2
retrying==1.3.4
rich==13.4.2
safetensors==0.4.2
scikit-image==0.24.0
scikit-learn==1.4.0
scipy==1.12.0
sentry-sdk==1.40.0
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
termcolor==2.4.0
thop==0.1.1.post2209072238
threadpoolctl==3.2.0
tifffile==2024.8.10
tokenizers==0.15.2
tomli==2.0.1
torch==2.2.0
torchvision==0.17.0
tqdm==4.65.2
transformers==4.38.2
triton==2.2.0
typing_extensions==4.9.0
tzdata==2024.1
urllib3==1.26.18
wandb==0.16.2
Werkzeug==3.0.2
yapf==0.40.2
zipp==3.18.1
